Project isolation and batching
Added shared project to allow better structure of code and added some
extranious validation to the HTTPTrigger input object.  Changed the
indexing function to work in batching mode to support large scale
updates.  This is the initial batching update, there's another SDK
capability that we need to explore that has inlligent batching and
indexing that is sent to Azure Search.