Paged results
Added supported for paging of data. Two new parameters are now required
for a QueryRequest -- PageSize and PageIndex. These control how data is
batched retrieved from Azure Search.  The next step will be to add
support and read up on OrderyBy to guarnatee stability of search
results. Based on docs a potential choice for search results is to sort
by Score. But we'll research to learn more about it