Paging (#4)
* Support for paging

Added code that leverages the $skip and $top to allow for paging of
results. This will guanratee we're not returning too many records and
triggering the RU limit for Cosmos.
Also switched the query results implementation from class to record
class to indicate it's an immutable object intended for returning to
requester, not for manipulation.